+++ test/manager.test.js
@@ -512,6 +512,50 @@ describe('Manager', function () {
         });
     });
 
+    describe('update', function() {
+
+        it('should update state', function(done) {
+            var name = 'iframe_update_state' + helpers.getRandomName();
+            var manager = helpers.testableManager();
+
+            var container = helpers.insertContainer(name);
+
+            manager.queue(name, {
+                'container': container,
+                'url': SCRIPT_URL,
+                'width': 123,
+                'height': 123,
+                'data': {another: 321}
+            });
+
+            manager.render(name, function(err, item){
+                expect(item.state).to.equal(State.RESOLVED);
+                expect(item.rendered.width).to.equal(123);
+                manager.update(name, {
+                    width: 300,
+                    height: 300,
+                    url: SCRIPT_URL + '?param1=23',
+                    data: {random: 123}
+                });
+
+                manager.refresh(name, function (err, item) {
+                    expect(item.options.data.random).to.equal(123);
+                    expect(item.options.data.another).to.equal(321);
+                    expect(item.options.width).to.equal(300);
+                    expect(item.options.height).to.equal(300);
+                    expect(item.options.url).to.equal(SCRIPT_URL + '?param1=23');
+
+                    expect(item.rendered.times).to.equal(2);
+                    expect(item.rendered.width).to.equal(300);
+                    expect(item.rendered.height).to.equal(300);
+                    done();
+                });
+
+            });
+        });
+
+    });
+
     describe('refresh', function () {
 
         it('should refresh single banner', function (done) {